Update 30/09/2013 WIP #2

+++ pyfltk_intro/musicplayer.py
@@ -1,5 +1,4 @@
 #!/usr/bin/python
-# -*- coding: utf-8 -*-
 
 import subprocess as subp
 
@@ -9,21 +8,20 @@ from fltk import *
 def playmusic(wid):
     mpath = "/home/inter/music/"
     fname = fl_file_chooser("Pick a music file", "*.ogg", mpath)
+    if fname is None:
+        return
     ps.append(subp.Popen(["ogg123", fname]))
     wid.deactivate()
-    pausebut.activate()
     stopbut.activate()
 
 
-def pausemusic(wid):
-    pass
-
-
 def stopmusic(wid):
-    ps.pop().kill()
-    playbut.activate()
-    pausebut.deactivate()
-    stopbut.deactivate()
+    if ps:
+        ps.pop().kill()
+        playbut.activate()
+        wid.deactivate()
+    if wid == win:
+        raise SystemExit, "\n\nGoodbye!"
 
 ps = []
 
@@ -33,14 +31,13 @@ win.begin()
 
 playbut = Fl_Button(10, 10, 40, 40, "@+4>")
 playbut.callback(playmusic)
-pausebut = Fl_Button(60, 10, 40, 40, "@circle")
-pausebut.callback(pausemusic)
-pausebut.deactivate()
 stopbut = Fl_Button(110, 10, 40, 40, "@square")
 stopbut.callback(stopmusic)
 stopbut.deactivate()
 
 win.end()
+
+win.callback(stopmusic)
 
 win.show()
 Fl.scheme("gtk+")
